e49d50cf40766120d4079f2ae6f5b33a5952c4d0 bench: fix 32-bit narrowing warning in bench/peer_eviction.cpp (Jon Atack)

Pull request description:

  Closes https://github.com/bitcoin/bitcoin/issues/22459.

ACKs for top commit:
  hebasto:
    ACK e49d50cf40766120d4079f2ae6f5b33a5952c4d0, tested on Debian 10.10 (i386):

Tree-SHA512: 8ca366fc296c633dbc8b8e0e7d80f4f6a64d02fb3da86d199881364f027d34b816a3c964b3fea2c1cc0b3ad51dd02d93c8bb14b5ebbd99fb4073cd1031766332

+++ src/bench/peer_eviction.cpp
@@ -27,7 +27,8 @@ static void EvictionProtectionCommon(
         candidate_setup_fn(c);
     }
 
-    std::vector<Candidates> copies{bench.epochs() * bench.epochIterations(), candidates};
+    std::vector<Candidates> copies{
+        static_cast<size_t>(bench.epochs() * bench.epochIterations()), candidates};
     size_t i{0};
     bench.run([&] {
         ProtectEvictionCandidatesByRatio(copies.at(i));
